# src/controllers/auth.js
import { ONE_MONTH } from '../constants.js';
import {
  loginOrSignUpWithGoogle,
  loginUser,
  logoutUser,
  refreshSession,
  registerUser,
  requestResetToken,
  resetPassword,
} from '../services/auth.js';
import { generateAuthUrl } from '../utils/googleOAuth2.js';

export async function registerUserController(req, res, next) {
  const newUser = await registerUser(req.body);

  res.status(201).json({
    status: 201,
    message: 'User registered successfully!',
    data: newUser,
  });
}

export async function loginUserController(req, res) {
  const session = await loginUser(req.body);

  await setupSession(res, session);

  res.status(200).json({
    status: 200,
    message: 'User is successfully logged in!',
    data: { accessToken: session.accessToken },
  });
}

export async function refreshSessionController(req, res) {
  const session = await refreshSession({
    sessionId: req.cookies.sessionId,
    refreshToken: req.cookies.refreshToken,
  });

  await setupSession(res, session);

  res.status(200).json({
    status: 200,
    message: 'Successfully refreshed a session!',
    data: { accessToken: session.accessToken },
  });
}

async function setupSession(res, session) {
  res.cookie('sessionId', session._id, {
    httpOnly: true,
    expires: new Date(Date.now() + ONE_MONTH),
  });
  res.cookie('refreshToken', session.refreshToken, {
    httpOnly: true,
    expires: new Date(Date.now() + ONE_MONTH),
  });
}

export async function logoutUserController(req, res) {
  if (req.cookies.sessionId) {
    await logoutUser(req.cookies.sessionId);
  }

  res.clearCookie('sessionId');
  res.clearCookie('refreshToken');

  res.status(204).send();
}

export async function requestResetPasswordController(req, res) {
  await requestResetToken(req.body.email);

  res.status(200).json({
    status: 200,
    message: 'Reset password email has been successfully sent.',
    data: {},
  });
}

export async function resetPasswordController(req, res) {
  await resetPassword(req.body);

  res.status(200).json({
    status: 200,
    message: 'Password has been successfully updated.',
    data: {},
  });
}

export async function getGoogleOAuthUrlController(req, res) {
  const url = generateAuthUrl();

  res.status(200).json({
    status: 200,
    message: 'Successfully sent auth link!',
    data: { url },
  });
}

export async function loginWithGoogleController(req, res) {
  const session = await loginOrSignUpWithGoogle(req.body.code);
  await setupSession(res, session);

  res.status(200).json({
    status: 200,
    message: 'Successfully logged in via Google OAuth!',
    data: { accessToken: session.accessToken },
  });
}